+++ pocketbase/migrations/0004_drop_canary.js
@@ -1,6 +1,0 @@
-migrate(
-  (app) => {
-    // no-op
-  },
-  (app) => {},
-)