+++ pocketbase/migrations/0005_drop_canary.js
@@ -1,9 +1,0 @@
-migrate(
-  (app) => {
-    try {
-      const col = app.findCollectionByNameOrId('_canary_check')
-      app.delete(col)
-    } catch (_) {}
-  },
-  (app) => {},
-)